Telekinetic Enemies and Enhanced Enemy Rounds
There are now enemies that can deflect bullets and thrash the player
around. The thrashing is disabled because it was extremely overpowered,
and could mean instant death in certain scenarios. I also added a wave
type where only enhanced enemies will spawn. This will be a part of the
special wave types that I want to try and implement.