Fixed Bullet Relative Velocity
For some reason, I didn't setup the bullets to move at a speed relative
to the shooter. So, I quickly fixed it, and based on a brief analysis, I
am pretty sure that the gun recoil isn't going to be needed. I also
amped up the speed limit and acceleration of the player.